fix(db): stitch session identity on db advisors --linked to match Go
Go wraps every Management API response in identityTransport, which fires
OnGotrueID -> StitchLogin on the X-Gotrue-Id header to persist the session's
authenticated identity (apps/cli-go/internal/utils/api.go:128,
identity_transport.go:12-20, cmd/root.go:146-154). The TS port only ran that
stitch inside the typed LegacyPlatformApi client's transport; db advisors
--linked deliberately uses a raw HttpClient (for enum tolerance) and omits that
client, so a session whose first API call is db advisors --linked silently
skipped identity stitching — invisible telemetry drift.

Hoist the stitch logic out of legacy-platform-api.layer.ts into a shared
legacy/shared/legacy-identity-stitch.ts (a factory the typed client keeps using
directly, plus a LegacyIdentityStitch service/layer for the raw path), and run
the stitch on each raw advisor response. One stitcher is shared across the
security + performance calls so it fires at most once per session, matching Go's
NeedsIdentityStitch gate.

Scoped to advisors; the other raw-HTTP Management API paths (config push,
postgres-config) have the same gap and are tracked as a follow-up.

Verified: apps/cli check:all, full unit (1220) + integration (1191) suites
(platform-api stitch behaviour unchanged), and a bundled-binary smoke test of
db advisors --linked (runtime constructs with the new layer, no missing-service
panic).
